+++ docs/changelog.rst
@@ -30,7 +30,8 @@ Changelog
 * :bug:`-` Compound finance COMP claim cases that were not decoded correctly should now appear fine.
 * :bug:`-` Fix an issue where some tokens weren't properly detected in scroll.
 * :feature:`-` Paladin bribe claiming events should now be properly decoded.
-* :bug:`-` Scroll decoders will work as expected now.
+* :bug:`-` Scroll decoders will now work as expected.
+* :bug:`-` Events claiming veCRV bribes from Gauge Bribe v2 will now be properly decoded.
 
 * :release:`1.33.1 <2024-05-29>`
 * :bug:`-` Fix the issue where airdrops aren't properly filtered by status.